ref: Move the jaas util out of specs/.
