+++ tests/specs/jaas/maxOccupants.spec.ts
@@ -1,5 +1,5 @@
 import { setTestProperties } from '../../helpers/TestProperties';
-import { joinMuc, generateJaasToken as t } from '../helpers/jaas';
+import { joinMuc, generateJaasToken as t } from '../../helpers/jaas';
 
 setTestProperties(__filename, {
     useJaas: true,
